+++ scrapegraphai/graphs/search_link_graph.py
@@ -1,18 +1,18 @@
-""" 
-SearchLinkGraph Module 
 """
+SearchLinkGraph Module
+"""
+
 from typing import Optional
 import logging
 from pydantic import BaseModel
 from .base_graph import BaseGraph
 from .abstract_graph import AbstractGraph
-from ..nodes import (FetchNode,
-                     SearchLinkNode,
-                     SearchLinksWithContext)
+from ..nodes import FetchNode, SearchLinkNode, SearchLinksWithContext
+
 
 class SearchLinkGraph(AbstractGraph):
-    """ 
-    SearchLinkGraph is a scraping pipeline that automates the process of 
+    """
+    SearchLinkGraph is a scraping pipeline that automates the process of
     extracting information from web pages using a natural language model
     to interpret and answer prompts.
 
@@ -22,7 +22,7 @@ class SearchLinkGraph(AbstractGraph):
         config (dict): Configuration parameters for the graph.
         schema (BaseModel): The schema for the graph output.
         llm_model: An instance of a language model client, configured for generating answers.
-        embedder_model: An instance of an embedding model client, 
+        embedder_model: An instance of an embedding model client,
         configured for generating embeddings.
         verbose (bool): A flag indicating whether to show print statements during execution.
         headless (bool): A flag indicating whether to run the graph in headless mode.
@@ -49,14 +49,15 @@ class SearchLinkGraph(AbstractGraph):
         """
 
         fetch_node = FetchNode(
-                input="url| local_dir",
-                output=["doc"],
-                node_config={
-                    "force": self.config.get("force", False),
-                    "cut": self.config.get("cut", True),
-                    "loader_kwargs": self.config.get("loader_kwargs", {}),
-                }
-            )
+            input="url| local_dir",
+            output=["doc"],
+            node_config={
+                "force": self.config.get("force", False),
+                "cut": self.config.get("cut", True),
+                "loader_kwargs": self.config.get("loader_kwargs", {}),
+                "storage_state": self.config.get("storage_state"),
+            },
+        )
 
         if self.config.get("llm_style") == (True, None):
             search_link_node = SearchLinksWithContext(
@@ -65,7 +66,7 @@ class SearchLinkGraph(AbstractGraph):
                 node_config={
                     "llm_model": self.llm_model,
                     "chunk_size": self.model_token,
-                }
+                },
             )
         else:
             search_link_node = SearchLinkNode(
@@ -73,19 +74,14 @@ class SearchLinkGraph(AbstractGraph):
                 output=["parsed_doc"],
                 node_config={
                     "chunk_size": self.model_token,
-                }
+                },
             )
 
         return BaseGraph(
-            nodes=[
-                fetch_node,
-                search_link_node
-            ],
-            edges=[
-                (fetch_node, search_link_node)
-            ],
+            nodes=[fetch_node, search_link_node],
+            edges=[(fetch_node, search_link_node)],
             entry_point=fetch_node,
-            graph_name=self.__class__.__name__
+            graph_name=self.__class__.__name__,
         )
 
     def run(self) -> str: